Throw an error at parse time for tags-with-spaces
To do so, we need a generic parser error; overload ParserSingle for the purpose.

We used to throw this error at pickle time, which is also weird, but parse time is even more wrong,
because we need to cache the error being thrown in order to be able to throw the same error on each
line check (which means we got our layering wrong... ah, well).