Further simplify the parser
Move the responsibility to check that there even *is* an input line to match,
to the token matcher. Now that the parser largely isn't responsible for it
anymore, we can do away with a whole slew of equally-named wrapper functions
in the parser that already exist in the token matcher. Not having two sets
of functions by the same name helps to navigate the code.