[php] Configure Psalm 6 default settings findUnusedBaselineEntry, findUnusedCode (#234)
* [php] Configure Psalm 6 default settings

This fixes a warning that outputs that we should be explicit about this setting. I picked enabling it as the solution.

* [php] Configure new Psalm checks OFF

We were not ready yet. These are now explicitly disabled, which takes care of the reminder message.

* [php] Psalm annotate exception class props.

This tells Psalm that these are a part of an expected API.

* Misc clean up

* Run `vendor/bin/psalm --no-cache --alter --issues=MissingParamType`
* Bump "phpunit/phpunit": "^10.5" to minimum deps run doesn't use a version that is too old
* Upgrade to latest phpunit.xml format
* Set findUnusedBaselineEntry="false" and findUnusedCode="false" to suppress paslm warnings

---------

Co-authored-by: M.P. Korstanje <rien.korstanje@gmail.com>