go: generate formatted parser
The go parser is currently generated and then formatted. This however
requires that go is installed when generating the parser. This
added dependency makes code generation more complicated than it needs
to be as this is yet another tool people need to install.

This in turn makes #154 more complicated than it needs to be.

By making the code generator generate formatted code we avoid this
problem all together.