java: Rewrite Groovy code generation to Java
Unfortunately the groovy-maven-plugin doesn't support Java 17 source
code yet (groovy/gmaven#28). While we currently only have to run on Java
17 for cucumber/cucumber-jvm#2644 it is not unimaginable that at some
point in the future we'll also want to compile our own source code with
17 as well.

With the increased release cadence of Java it would be prudent to keep
the technology stack shallow and remove the dependency on Groovy and
Groovy related components al together. So replacing the whole stack with
pure Java would be beneficial.

Fixes: #90