Prefer the longest step keyword
Step keywords may be prefixes of each other. For example in French
several variations of given start with `Etant donné`. When this is the
case Gherkin should prefer the longest possible keyword.

This is currently only implicitly enforced by manually ordering the
keywords.

However, for Creole language the prefixes are spread between then and
when so manually sorting them is no longer possible.