go: Prefer the longest keywords (#404)
Some langauges have two different step keywords starting with the same
characters. In cases like these, we need to ensure that if a user types
out the longer of these keywords, that is what gets matched.

This is done by sorting the keywords for each dialect.

As well as this, I've added handling to make sure that the sorting is
only done once per dialect to make it inline with other fixes. This
required a couple of fields to be added to the `Dialect` struct, so I've
updated the `builtin_dialects.go.jq` file to built the `Dialect` struct
in a way that allows zero base instantiation.