Fix empty token location (#440)
While investigating the impact of #436 and #438 some discrepancies came
to light.

Lines consisting entirely of whitespace are considered empty. As such
the tokens for these lines should have 0 indentation and start at column
1.

This behavior diverged for c, .net, go, java and php. As tokens are not
part of the public API, this change has no impact on the observable
behavior of the parser.