Improve parser performance (#436)
We tracked down every millisecond, and found several improvements, so
that in the end, small streams make big rivers. But that means that
sometimes we change the code to gain just a few milliseconds, so not
everyone may like it and take the risk to change the code.

Some example of optimization techniques used:

 * avoid dynamic lambda creation
 * precomputing String length
 * replacing String instantiation by String char-by-char examination
 * specific treatment depending on use-case (one-letter separator, empty collection, pre-sized collection, ...)
 * use of specific interface to pass parameters (Supplier -> BiConsumer/BiPredicate)
 * replaced for-each by classical for

Co-authored-by: Julien Kronegg <julien@kronegg.ch>
Co-authored-by: M.P. Korstanje <rien.korstanje@gmail.com>