(dotnet) Migrate gherkin messages to cucumber messages (#426)
This PR enhances the .Net implementation such that the Message events
and pickles that are generated use the Cucumber/Messages types instead
of the Gherkin.Cucumber.Messages.Types which had been embedded in the
Gherkin project.


* Add test data for star keywords

The starkey word is kinda special because its step type is unknown.
How different implementations handle this seems to be inconsistent. This
should highlight some problems.

Co-authored-by: M.P. Korstanje <rien.korstanje@gmail.com>